+++ src/test/scala/ChatSpec.scala
@@ -1,5 +1,5 @@
-import com.flyberrycapital.slack.HttpClient
-import com.flyberrycapital.slack.Methods.Chat
+import com.ponkotuy.slack.HttpClient
+import com.ponkotuy.slack.Methods.Chat
 import org.joda.time.{DateTime, DateTimeZone}
 import org.mockito.Mockito._
 import org.scalatest.mock.MockitoSugar
